refactor: completelty refactor detail view structure
