+++ project-canvas/src/components/DetailView/Components/AddSubtask/queries.ts
@@ -1,0 +1,22 @@
+import { showNotification } from "@mantine/notifications"
+import { QueryClient, useMutation } from "@tanstack/react-query"
+import { createSubtask } from "./queryFunctions"
+
+export const createSubtaskMutation = (
+  issueKey: string,
+  summary: string,
+  projectId: string,
+  queryClient: QueryClient,
+  resetSummary: () => void
+) =>
+  useMutation({
+    mutationFn: () => createSubtask(issueKey, summary, projectId),
+    onSuccess(createdSubtask: { id: string; key: string }) {
+      resetSummary()
+      showNotification({
+        message: `The issue  ${createdSubtask.key} has been created!`,
+        color: "green",
+      })
+      queryClient.invalidateQueries({ queryKey: ["issues"] })
+    },
+  })
