+++ project-canvas/src/components/DetailView/Components/AssigneeMenu/index.ts
@@ -1,0 +1,1 @@
+export * from "./AssigneeMenu"
