+++ project-canvas/src/components/DetailView/Components/AssigneeMenu/queryFunctions.ts
@@ -1,0 +1,12 @@
+import { User } from "project-extender"
+
+export const getAssignableUsersByProject = (
+  projectIdOrKey: string
+): Promise<User[]> =>
+  fetch(
+    `${
+      import.meta.env.VITE_EXTENDER
+    }/assignableUsersByProject?projectIdOrKey=${projectIdOrKey}`
+  )
+    .then((users) => users.json())
+    .catch((err) => err)
