+++ project-canvas/src/components/DetailView/Components/CommentSection/index.ts
@@ -1,0 +1,1 @@
+export * from "./CommentSection"
