+++ project-canvas/src/components/DetailView/Components/EditableEpic/index.ts
@@ -1,0 +1,1 @@
+export * from "./EditableEpic"
