+++ project-canvas/src/components/DetailView/Components/EditableEpic/queryFunctions.ts
@@ -1,0 +1,10 @@
+import { Issue } from "project-extender"
+
+export const getEpicsByProject = (projectIdOrKey: string): Promise<Issue[]> =>
+  fetch(
+    `${
+      import.meta.env.VITE_EXTENDER
+    }/epicsByProject?projectIdOrKey=${projectIdOrKey}`
+  )
+    .then((epics) => epics.json())
+    .catch((err) => err)
