+++ project-canvas/src/components/DetailView/Components/IssueSummary.tsx
@@ -3,7 +3,7 @@ import { Issue } from "project-extender"
 import { useState } from "react"
 import { showNotification } from "@mantine/notifications"
 import { useMutation } from "@tanstack/react-query"
-import { editIssue } from "../../CreateIssue/queryFunctions"
+import { editIssue } from "../helpers/queryFunctions"
 
 export function IssueSummary({
   summary,
