+++ project-canvas/src/components/DetailView/Components/StoryPointsEstimateMenu.tsx
@@ -3,10 +3,8 @@ import { Text, Group, NumberInput, Chip, Loader, Box } from "@mantine/core"
 import { showNotification } from "@mantine/notifications"
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query"
 import { Issue } from "project-extender"
-import {
-  editIssue,
-  getEditableIssueFields,
-} from "../../CreateIssue/queryFunctions"
+import { getEditableIssueFields } from "../../CreateIssue/queryFunctions"
+import { editIssue } from "../helpers/queryFunctions"
 
 export function StoryPointsEstimateMenu({
   issueKey,
