+++ project-canvas/src/components/DetailView/Components/SubTask/index.ts
@@ -1,0 +1,1 @@
+export * from "./Subtask"
